requirements.txt
# Core Installation
-e .   # Installs your package and core dependencies from setup.py

# Development Tools
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
pytest>=7.0.0

# Explicit listing of all dependencies (including those in setup.py)
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
torchmetrics>=1.0.0
scipy>=1.10.0
statistics>=1.0.0
tokenizers>=0.13.0
wandb>=0.15.0
tqdm>=4.65.0
psutil>=5.9.0
python-dotenv>=1.0.0
pyyaml>=6.0
python-box>=7.0.1
ensure>=1.0.3
joblib>=1.3.0
pathlib>=1.0.1
scikit-learn>=1.0.0
plotly>=5.0.0
jupyter>=1.0.0